Support slow player initialization through delayed retries
Drop unnecessary encapsulation in self-executing function

Content scripts run in a separate JS context.
In Firefox this is called ‘Xray vision’.
The content script has no access or visibility to the website JS context, or vice versa.
Hence, there is no need to secure the script against conflicts and for encapsulation.

https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Content_scripts#dom_access